Exercise 2 Arithmetic Operators
Use the arithmetic operators in the manner and order in which they are listed below.
For those who feel inclined, try challenging yourself by completing all operations in
one statement remembering the rules of PEMDAS.
Use the multiplication operator  *  to double  the value of the variable  number and store the result back in number .
Use the addition operator  +  to add 9  to the variable  number  and store the result back in number .
Use the subtraction operator  -  to subtract 3  from the variable number  and store the result back in number .new value.
Use the division operator  /  to divide the the variable number  by 2   and store the result back in number .
Use the subtraction operator  -  to subtract the variable named original_number  from the variable number  and store the result back in number .
Use the modulo operator  %  to find the remainder of the new value when divided by 3  and store the result back in number .