Challenge
The idea here is to write a program that asks the user to enter the following:
An integer representing the number of cents.
You can assume that the number of cents entered is greater than or equal to zero.
The program then displays how to provide that change to the user.
In the U.S., one dollar is 100 cents, a quarter is 25 cents, a dime is 10 cents, a nickel is 5 cents, and a penny is one cent.
Do this using modulus and not using modulus.