Exercise 4 Switch Statement
Switch Statement - Day of the Week
In this exercise you will write a switch statement that displays the name of the day of the week given the day code.
You may use the integer variable named day_code whose value we will change to automatically to test your code.
Given the following day codes, your program should display the day of the week using a cout statement.