Exercise 9 Nested Loop
Nested Loops - Sum of the Product of all Pairs of Vector Elements
Given a vector of integers named vec  that is provided for you,
find the sum of the product of all pairs of vector elements.
You should declare an integer variable named result
and store the final product in this variable.