Challenge
Challenge Requirements
You need to write a program that:
Asks the user to enter a secret message (a string).
Encrypts that message using the substitution cipher.
Displays the encrypted message.
Then decrypts the encrypted message back to the original message.
You can use two strings for your substitution: one string for the alphabet and another
string for the key.
A simple and very old method of sending secret messages is the substitution cipher.
You might have used this cipher with your friends when you were a kid.
Basically, each letter of the alphabet gets replaced by another letter of the alphabet.
For example, every 'a' get replaced with an 'X', and every 'b' gets replaced with a 'Z', etc.
Write a program that asks a user to enter a secret message.
Encrypt this message using the substitution cipher and display the encrypted message.
Then decrypt the encrypted message back to the original message.
You may use the 2 strings below for your substitution.
For example, to encrypt you can replace the character at position n in alphabet
with the character at position n in key.
To decrypt you can replace the character at position n in key
with the character at position n in alphabet.
Remember, the less code you write the less code you have to test!
Reuse existing functionality in libraries and in the std::string class!