Exercise 6 Pass by Reference
Using Pass by Reference - Print a Guest List
In this exercise you will rewrite the previous Guest List exercise only this time
with the use of reference variables.

In this exercise you will create a program that will be used to print the guest list
to an event and then clear the guest list when the event is over.