Exercise 7 Recursive Function
Implementing a Recursive Function - Sum of Digits
Problem Title: Sum of Digits using Recursion

Problem Description:
Write a C++ program to find the sum of digits of a number using recursion.
You need to implement the function int sum_of_digits(int n) that takes an
integer n as an argument and returns the sum of its digits.