Exercise_1 Classes and Objects
Creating and Accessing Objects
In the test_dog function create a Dog  object named spot and, using the . operator, set the name  attribute to the string "Spot"  and the age  attribute to 5 .

Adding public methods that access private class attributes:
get_name returns the name attribute of the dog
set_name sets the name attribute of the dog to the string that is passed into the method.
get_age returns the age attribute of the dog
set_age sets the age attribute of the dog to the int passed into the method.
get_human_years() return the dog's age * 7
speak()  returns the std::string "Woof"